MP-160 - refetch completeness data when user is updated
