+++ src/lib/functions/user-profile.provider.ts
@@ -55,46 +55,3 @@ export const fetchUserProfile = async (): Promise<AuthUser> => {
 
   return localCache[userHandle];
 }
-
-export interface ProfileCompletednessResponse {
-  handle: string
-  showToast: string
-  data: {
-    percentComplete: number
-  }
-}
-
-/**
- * Fetches the user profile completedness
- * @returns Promise<ProfileCompletednessResponse>
- */
-export const fetchUserProfileCompletedness = async (user: AuthUser): Promise<ProfileCompletednessResponse> => {
-  const userHandle = user?.handle;
-
-  if (!userHandle) {
-    return;
-  }
-
-  const cacheKey = `${userHandle}-completedness`;
-  if (localCache[cacheKey]) {
-    return await localCache[cacheKey];
-  }
-
-
-  let resolve: (value: ProfileCompletednessResponse) => void;
-  localCache[cacheKey] = new Promise((r) => {resolve = r});
-
-  const requestUrl: string = `${TC_API_V5_HOST}/members/${userHandle}/profileCompleteness`;
-  const request = fetch(requestUrl, {headers: {...getRequestAuthHeaders()}});
-
-  const response = await (await request).json();
-  resolve({
-    ...response,
-    data: {
-      ...response.data,
-      percentComplete: (response?.data?.percentComplete ?? 0) * 100,
-    },
-  });
-
-  return localCache[cacheKey];
-}
